Update dd bs test scripts with version tested on OSX Yosemite

+++ lib/blog_snippets/tuning_dd_block_size/dd_obs_test.sh
@@ -1,17 +1,33 @@
 #!/bin/bash
 
-TEST_FILE=${1:-dd_obs_testfile}
-TEST_FILE_SIZE=128M
+# Since we're dealing with dd, abort if any errors occur
+set -e
 
-for block_size in 512 1K 2K 4K 8K 16K 32K 64K 128K 256K 512K 1M 2M 4M 8M 16M 32M 64M
+TEST_FILE=${1:-dd_obs_testfile}
+TEST_FILE_SIZE=134217728
+
+# Header
+PRINTF_FORMAT="%8s : %s\n"
+printf "$PRINTF_FORMAT" 'block size' 'transfer rate'
+
+# Block sizes of 512b 1K 2K 4K 8K 16K 32K 64K 128K 256K 512K 1M 2M 4M 8M 16M 32M 64M
+for BLOCK_SIZE in 512 1024 2048 4096 8192 16384 32768 65536 131072 262144 524288 1048576 2097152 4194304 8388608 16777216 33554432 67108864
 do
+  # Calculate number of segments required to copy
+  COUNT=$(($TEST_FILE_SIZE / $BLOCK_SIZE))
+
+  if [ $COUNT -le 0 ]; then
+    echo "Block size of $BLOCK_SIZE estimated to require $COUNT blocks, aborting further tests."
+    break
+  fi
+
   # Create a test file with the specified block size
-  dd_result=$(dd if=/dev/zero of=$TEST_FILE iflag=count_bytes bs=$block_size count=$TEST_FILE_SIZE 2>&1 1>/dev/null)
+  DD_RESULT=$(dd if=/dev/zero of=$TEST_FILE bs=$BLOCK_SIZE count=$COUNT 2>&1 1>/dev/null)
 
   # Extract the transfer rate from dd's STDERR output
-  transfer_rate=$(echo $dd_result | \grep --only-matching -E '[0-9.]+ [MGk]?B/s')
+  TRANSFER_RATE=$(echo $DD_RESULT | \grep --only-matching -E '[0-9.]+ ([MGk]?B|bytes)/s(ec)?')
 
   # Clean up the test file and output result
   rm $TEST_FILE
-  echo "$block_size: $transfer_rate"
+  printf "$PRINTF_FORMAT" "$BLOCK_SIZE" "$TRANSFER_RATE"
 done
